+++ fbw-a380x/src/systems/instruments/src/MFD/FMC/FlightManagementComputer.ts
@@ -1642,6 +1642,7 @@ export class FlightManagementComputer implements FmcInterface {
     const activePlan = this.flightPlanInterface.active;
     activePlan.performanceData.preselectedClimbSpeed.set(null);
     activePlan.performanceData.preselectedCruiseSpeed.set(null);
+    let stepDeleted = false;
 
     // Delete planned/future altitude steps
     this.flightPlanInterface.active.allLegs
@@ -1652,8 +1653,13 @@ export class FlightManagementComputer implements FmcInterface {
       .forEach((l) => {
         if (l.isDiscontinuity === false) {
           l.cruiseStep = undefined;
+          stepDeleted = true;
         }
       });
+
+    if (stepDeleted) {
+      this.addMessageToQueue(NXSystemMessages.stepDeleted);
+    }
 
     // Delete time constraints
     // no-op
@@ -1664,6 +1670,10 @@ export class FlightManagementComputer implements FmcInterface {
 
   public exitEngineOut() {
     // Restore long-term guidance targets
+  }
+
+  public engineOutActive(): boolean {
+    return this.fmgc.data.engineOut.get();
   }
 
   async swapNavDatabase(): Promise<void> {