+++ fbw-a380x/src/systems/instruments/src/MFD/pages/FMS/F-PLN/MfdFmsFplnVertRev.tsx
@@ -29,14 +29,14 @@ import {
 } from 'instruments/src/MFD/pages/common/DataEntryFormats';
 import { DropdownMenu } from 'instruments/src/MsfsAvionicsCommon/UiWidgets/DropdownMenu';
 import { Vmo } from '@shared/PerformanceConstants';
-import { FlightPlanLeg } from '@fmgc/flightplanning/legs/FlightPlanLeg';
+import { isLeg } from '@fmgc/flightplanning/legs/FlightPlanLeg';
 import { RadioButtonColor, RadioButtonGroup } from 'instruments/src/MsfsAvionicsCommon/UiWidgets/RadioButtonGroup';
 import { AltitudeDescriptor, WaypointConstraintType } from '@flybywiresim/fbw-sdk';
 import { IconButton } from 'instruments/src/MsfsAvionicsCommon/UiWidgets/IconButton';
 import { FmgcData } from 'instruments/src/MFD/FMC/fmgc';
 import { CruiseStepEntry } from '@fmgc/flightplanning/CruiseStep';
 import { NXSystemMessages } from 'instruments/src/MFD/shared/NXSystemMessages';
-import { getEtaFromUtcOrPresent } from 'instruments/src/MFD/shared/utils';
+import { getEtaFromUtcOrPresent, isConstraintRevisionAllowed } from 'instruments/src/MFD/shared/utils';
 import { FmgcFlightPhase } from '@shared/flightphase';
 import { ReadonlyFlightPlan } from '@fmgc/flightplanning/plans/ReadonlyFlightPlan';
 import { ReadonlyFlightPlanLeg } from '@fmgc/flightplanning/legs/ReadonlyFlightPlanLeg';
@@ -57,32 +57,32 @@ enum SelectedPage {
   STEP_ALTS = 4,
 }
 
+enum StepDisabledReason {
+  EngineOut = 'STEP ALTs NOT ALLOWED IN EO',
+  Descent = 'STEP ALTs NOT ALLOWED IN DES',
+  Approach = 'STEP ALTs NOT ALLOWED IN APPR',
+  GoAround = 'STEP ALTs NOT ALLOWED IN GA',
+  Done = 'STEP ALTs NOT ALLOWED IN DONE',
+  NoCruiseLevel = 'STEP ALTs NOT ALLOWED: NO CRZ FL',
+}
+
 export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
   private readonly selectedPageIndex = Subject.create(SelectedPage.RTA);
-
+  private availableWaypointsIdents: string[] = [];
   private readonly availableWaypoints = ArraySubject.create<string>([]);
+  private initialLoadRevisedWaypointIndex = true;
 
   private availableWaypointsToLegIndex: number[] = [];
-
-  /** Don't modify this directly, use selectedLegIndex for conversion to/from leg indices */
-  private readonly dropdownMenuSelectedWaypointIndex = Subject.create<number | null>(null);
-
-  get selectedLegIndex() {
-    const dropdownIndex = this.dropdownMenuSelectedWaypointIndex.get();
-    return dropdownIndex !== null && dropdownIndex in this.availableWaypointsToLegIndex
-      ? this.availableWaypointsToLegIndex[dropdownIndex]
-      : null;
-  }
-
-  set selectedLegIndex(legIndex: number | null) {
-    if (legIndex === null) {
-      this.dropdownMenuSelectedWaypointIndex.set(null);
-      return;
+  private readonly selectedLegIndex = Subject.create<number | null>(null);
+  private readonly selectedLegIsAlternate = Subject.create<boolean | null>(null);
+  private readonly dropdownMenuSelectedWaypointIndex = this.selectedLegIndex.map((si) => {
+    if (si === null) {
+      return null;
+    } else {
+      const idx = this.availableWaypointsToLegIndex.findIndex((i) => i === si);
+      return idx !== -1 ? idx : null;
     }
-
-    const dropdownIndex = this.availableWaypointsToLegIndex.indexOf(legIndex);
-    this.dropdownMenuSelectedWaypointIndex.set(dropdownIndex !== -1 ? dropdownIndex : null);
-  }
+  });
 
   private readonly returnButtonDiv = FSComponent.createRef<HTMLDivElement>();
 
@@ -152,7 +152,7 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
 
   private readonly altConstraintTransitionIsFlightLevel = Subject.create<boolean>(false);
 
-  /** 0: CLB, 1: DES */
+  /** 0: AT, 1: AT OR ABOVE 2: AT OR BELOW */
   private readonly selectedAltitudeConstraintOption = Subject.create<number | null>(null);
 
   private readonly selectedAltitudeConstraintDisabled = this.altConstraintDisabled.map((it) => Array(3).fill(it));
@@ -172,6 +172,18 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
   // STEP ALTs page
   /** Paused when not in STEP ALts page */
   private stepsAltsClockSub?: Subscription;
+
+  private readonly stepPageDisabledReason = Subject.create<StepDisabledReason | null>(null);
+  private readonly stepPageDisabledMessageDisplay = this.stepPageDisabledReason.map((reason) =>
+    reason ? 'inherit' : 'none',
+  );
+  private readonly stepAltsPageDisplay = this.stepPageDisabledReason.map((reason) => (reason ? 'none' : 'block'));
+  private readonly fillPageDisplay = MappedSubject.create(
+    ([page, stepAltDisabled]) => (page === SelectedPage.STEP_ALTS && !stepAltDisabled ? 'inherit' : 'none'),
+    this.selectedPageIndex,
+    this.stepPageDisabledReason,
+  );
+
   private readonly crzFl = Subject.create<number | null>(null);
   private readonly crzFlFormatted = FmgcData.fmcFormatValue(this.crzFl);
 
@@ -222,31 +234,41 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
     this.transitionAltitude.set(pd?.transitionAltitude.get() ?? null);
     this.transitionLevel.set(pd?.transitionLevel.get() ?? null);
 
-    const activeLegIndex = this.props.flightPlanInterface.get(this.loadedFlightPlanIndex.get()).activeLegIndex;
-    if (activeLegIndex) {
-      const selectedLegIndexBeforeUpdate = this.selectedLegIndex;
-      this.availableWaypointsToLegIndex = [];
-      const wpt = this.loadedFlightPlan?.allLegs
-        .slice(activeLegIndex)
-        .map((el, idx) => {
-          if (el instanceof FlightPlanLeg && el.isXF()) {
-            this.availableWaypointsToLegIndex.push(idx + activeLegIndex);
-            return el.ident;
-          }
-          return null;
-        })
-        .filter((el) => el !== null) as string[] | undefined;
-      if (wpt) {
-        this.availableWaypoints.set(wpt);
-        this.selectedLegIndex = selectedLegIndexBeforeUpdate;
-      }
+    // Do not update till a flightplan has been fully loaded.
+    const fpLoaded = this.selectedLegIsAlternate.get() !== null;
+    if (!fpLoaded) {
+      return;
+    }
 
-      const revWptIdx = this.props.fmcService.master.revisedLegIndex.get();
-      this.selectedLegIndex = revWptIdx ?? null;
+    const plan = this.selectedLegIsAlternate.get() ? this.loadedAlternateFlightPlan : this.loadedFlightPlan;
+    const activeLegIndex = plan?.activeLegIndex;
+    let indexToSelect: number | null = null;
+    this.availableWaypointsToLegIndex = [];
+    this.availableWaypointsIdents = [];
+    if (activeLegIndex !== undefined && plan !== null) {
+      for (let i = activeLegIndex; i < plan.legCount; i++) {
+        const leg = plan.maybeElementAt(i);
+        if (isLeg(leg) && leg.isXF()) {
+          this.availableWaypointsToLegIndex.push(i);
+          this.availableWaypointsIdents.push(leg.ident);
+        }
+      }
+      indexToSelect = this.availableWaypointsToLegIndex[0] ?? null;
+    }
+    this.availableWaypoints.set(this.availableWaypointsIdents);
+    if (this.initialLoadRevisedWaypointIndex) {
+      this.selectedLegIndex.set(this.props.fmcService.master.revisedLegIndex.get() ?? indexToSelect);
+      this.initialLoadRevisedWaypointIndex = false;
+    } else {
+      const currentSelection = this.selectedLegIndex.get();
+      this.selectedLegIndex.set(
+        currentSelection !== null && this.availableWaypointsToLegIndex.includes(currentSelection)
+          ? currentSelection
+          : indexToSelect,
+      );
     }
 
     this.crzFl.set(pd?.cruiseFlightLevel.get() ?? null);
-
     this.updateConstraints();
     this.updateCruiseSteps();
   }
@@ -256,28 +278,37 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
     leg: ReadonlyFlightPlanLeg,
     flightPlan: ReadonlyFlightPlan,
   ): boolean {
-    // Check conditions: No constraints for airports, FROM waypoint, GA legs, pseudo waypoints
-    return (
-      !leg.isRunway() && legIndex >= flightPlan.activeLegIndex && legIndex < flightPlan.firstMissedApproachLegIndex
-    );
+    return isConstraintRevisionAllowed(leg) && legIndex >= flightPlan.activeLegIndex;
   }
 
   private updateConstraints() {
-    if (!this.props.fmcService.master || !this.loadedFlightPlan || this.selectedLegIndex === null) {
+    if (!this.props.fmcService.master || this.selectedLegIsAlternate.get() === null) {
       return;
     }
 
-    const leg = this.loadedFlightPlan.legElementAt(this.selectedLegIndex);
-    const previousElement = this.loadedFlightPlan.maybeElementAt(this.selectedLegIndex - 1);
-    const isPartOfTooSteepPathSegment =
-      leg.calculated?.endsInTooSteepPath ||
-      (previousElement?.isDiscontinuity === false && previousElement.calculated?.endsInTooSteepPath);
+    const plan = this.selectedLegIsAlternate.get() ? this.loadedAlternateFlightPlan : this.loadedFlightPlan;
+    const selectedLegIdx = this.selectedLegIndex.get();
+    const leg = selectedLegIdx !== null ? plan?.legElementAt(selectedLegIdx) : null;
+    const previousElement = selectedLegIdx !== null ? plan?.maybeElementAt(selectedLegIdx - 1) : null;
+    const isPartOfTooSteepPathSegment = leg
+      ? leg.calculated?.endsInTooSteepPath ||
+        (previousElement?.isDiscontinuity === false && previousElement.calculated?.endsInTooSteepPath)
+      : null;
 
-    if (!MfdFmsFplnVertRev.isEligibleForVerticalRevision(this.selectedLegIndex, leg, this.loadedFlightPlan)) {
+    if (
+      plan &&
+      leg &&
+      selectedLegIdx !== null &&
+      !MfdFmsFplnVertRev.isEligibleForVerticalRevision(selectedLegIdx, leg, plan)
+    ) {
       this.speedMessageArea.set(`SPD CSTR NOT ALLOWED AT ${leg.ident}`);
       this.spdConstraintDisabled.set(true);
+      this.spdConstraintTypeRadioSelected.set(null);
+      this.speedConstraintInput.set(null);
       this.altitudeMessageArea.set(`ALT CSTR NOT ALLOWED AT ${leg.ident}`);
       this.altConstraintDisabled.set(true);
+      this.altitudeConstraintInput.set(null);
+      this.selectedAltitudeConstraintOption.set(null);
       return;
     }
     this.speedMessageArea.set('');
@@ -288,9 +319,16 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
     // Load speed constraints
     const selectedIndex = this.selectedPageIndex.get();
     if (selectedIndex === SelectedPage.SPD) {
+      const constraintType =
+        plan && selectedLegIdx !== null && leg?.constraintType === WaypointConstraintType.Unknown
+          ? plan.autoConstraintTypeForLegIndex(selectedLegIdx)
+          : leg?.constraintType ?? WaypointConstraintType.Unknown;
       const speedLimitType =
-        leg.constraintType === WaypointConstraintType.DES ||
-        (this.activeFlightPhase.get() > FmgcFlightPhase.Cruise &&
+        constraintType === WaypointConstraintType.DES ||
+        (this.props.fmcService.master.flightPlanInterface
+          .get(this.loadedFlightPlanIndex.get())
+          .isActiveOrCopiedFromActive() &&
+          this.activeFlightPhase.get() > FmgcFlightPhase.Cruise &&
           this.activeFlightPhase.get() < FmgcFlightPhase.GoAround)
           ? SpeedLimitType.DES
           : SpeedLimitType.CLB;
@@ -301,76 +339,95 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
         speedLimitType === SpeedLimitType.CLB ? this.transitionAltitude.get() : this.transitionLevel.get(),
       );
 
-      // TODO handle alternate?
       const climbSpeedLimit = speedLimitType === SpeedLimitType.CLB;
+      const isAlternate = this.selectedLegIsAlternate.get();
 
       const speedLimitSpeed = climbSpeedLimit
-        ? this.loadedFlightPlan.performanceData.climbSpeedLimitSpeed.get()
-        : this.loadedFlightPlan.performanceData.descentSpeedLimitSpeed.get();
+        ? isAlternate
+          ? plan?.performanceData.alternateClimbSpeedLimitSpeed.get()
+          : plan?.performanceData.climbSpeedLimitSpeed.get()
+        : isAlternate
+          ? plan?.performanceData.alternateDescentSpeedLimitSpeed.get()
+          : plan?.performanceData.descentSpeedLimitSpeed.get();
 
       const speedLimitAltitude = climbSpeedLimit
-        ? this.loadedFlightPlan.performanceData.climbSpeedLimitAltitude.get()
-        : this.loadedFlightPlan.performanceData.descentSpeedLimitAltitude.get();
+        ? isAlternate
+          ? plan?.performanceData.alternateClimbSpeedLimitAltitude.get()
+          : plan?.performanceData.climbSpeedLimitAltitude.get()
+        : isAlternate
+          ? plan?.performanceData.alternateDescentSpeedLimitAltitude.get()
+          : plan?.performanceData.descentSpeedLimitAltitude.get();
 
-      this.speedLimitSpeed.set(speedLimitSpeed);
-      this.speedLimitAltitude.set(speedLimitAltitude);
+      this.speedLimitSpeed.set(speedLimitSpeed ?? null);
+      this.speedLimitAltitude.set(speedLimitAltitude ?? null);
       this.speedLimitPilotEntered.set(
         climbSpeedLimit
-          ? this.loadedFlightPlan.performanceData.isClimbSpeedLimitPilotEntered.get()
-          : this.loadedFlightPlan.performanceData.isDescentSpeedLimitPilotEntered.get(),
+          ? isAlternate
+            ? plan?.performanceData.isAlternateClimbSpeedLimitPilotEntered.get() ?? false
+            : plan?.performanceData.isClimbSpeedLimitPilotEntered.get() ?? false
+          : isAlternate
+            ? plan?.performanceData.isAlternateDescentSpeedLimitPilotEntered.get() ?? false
+            : plan?.performanceData.isDescentSpeedLimitPilotEntered.get() ?? false,
       );
 
-      this.speedConstraintInput.set(leg.speedConstraint?.speed ?? null);
+      this.speedConstraintInput.set(leg?.speedConstraint?.speed ?? null);
       this.constraintType.set(
-        leg.constraintType === WaypointConstraintType.CLB
+        constraintType === WaypointConstraintType.CLB
           ? 'CLB'
-          : leg.constraintType === WaypointConstraintType.DES
+          : constraintType === WaypointConstraintType.DES
             ? 'DES'
             : '',
       );
       this.spdConstraintTypeRadioSelected.set(
-        leg.constraintType === WaypointConstraintType.CLB
-          ? 0
-          : leg.constraintType === WaypointConstraintType.DES
-            ? 1
-            : null,
+        constraintType === WaypointConstraintType.CLB ? 0 : constraintType === WaypointConstraintType.DES ? 1 : null,
       );
+      this.cannotDeleteSpeedConstraint.set(!leg?.speedConstraint || !leg?.speedConstraint?.speed);
     } else if (selectedIndex === SelectedPage.ALT) {
+      const constraintType =
+        plan && selectedLegIdx !== null && leg?.constraintType === WaypointConstraintType.Unknown
+          ? plan.autoConstraintTypeForLegIndex(selectedLegIdx)
+          : leg?.constraintType ?? WaypointConstraintType.Unknown;
+      this.constraintType.set(
+        constraintType === WaypointConstraintType.CLB
+          ? 'CLB'
+          : constraintType === WaypointConstraintType.DES
+            ? 'DES'
+            : '',
+      );
       this.altConstraintTypeRadioSelected.set(
-        leg.constraintType === WaypointConstraintType.CLB
-          ? 0
-          : leg.constraintType === WaypointConstraintType.DES
-            ? 1
-            : null,
+        constraintType === WaypointConstraintType.CLB ? 0 : constraintType === WaypointConstraintType.DES ? 1 : null,
       );
       this.altitudeClbDesConstraintVisibility.set(
-        leg.constraintType === WaypointConstraintType.Unknown ? 'visible' : 'hidden',
+        constraintType === WaypointConstraintType.Unknown ? 'visible' : 'hidden',
       );
 
       this.cannotDeleteAltConstraint.set(
-        !leg.altitudeConstraint || leg.altitudeConstraint?.altitudeDescriptor === AltitudeDescriptor.None,
+        !leg?.altitudeConstraint || leg.altitudeConstraint?.altitudeDescriptor === AltitudeDescriptor.None,
       );
-      this.cannotDeleteSpeedConstraint.set(!leg.speedConstraint || !leg.speedConstraint?.speed);
 
       // Load altitude constraints
-      // FIXME missing a lot of cases here
-      switch (leg.altitudeConstraint?.altitudeDescriptor) {
+      switch (leg?.altitudeConstraint?.altitudeDescriptor) {
         case AltitudeDescriptor.AtAlt1:
+        case AltitudeDescriptor.AtAlt1GsIntcptAlt2:
+        case AltitudeDescriptor.AtAlt1AngleAlt2:
           this.selectedAltitudeConstraintOption.set(0);
           break;
         case AltitudeDescriptor.AtOrAboveAlt1:
+        case AltitudeDescriptor.AtOrAboveAlt1GsIntcptAlt2:
+        case AltitudeDescriptor.AtOrAboveAlt1AngleAlt2:
           this.selectedAltitudeConstraintOption.set(1);
           break;
         case AltitudeDescriptor.AtOrBelowAlt1:
+        case AltitudeDescriptor.AtOrBelowAlt1AngleAlt2:
           this.selectedAltitudeConstraintOption.set(2);
           break;
         default:
           this.selectedAltitudeConstraintOption.set(null);
           break;
       }
-      this.altitudeConstraintInput.set(leg.altitudeConstraint?.altitude1 ?? null);
+      this.altitudeConstraintInput.set(leg?.altitudeConstraint?.altitude1 ?? null);
 
-      const ac = leg.altitudeConstraint;
+      const ac = leg?.altitudeConstraint;
       if (ac) {
         const transAltIsFlightLevel = leg.constraintType === WaypointConstraintType.DES;
         const transAlt = transAltIsFlightLevel ? this.transitionLevel.get() : this.transitionAltitude.get();
@@ -431,95 +488,127 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
   }
 
   private updateCruiseSteps() {
-    if (this.loadedFlightPlan && this.selectedPageIndex.get() === SelectedPage.STEP_ALTS) {
-      const activeLegIndex = this.loadedFlightPlan.activeLegIndex;
-      const cruiseSteps = this.loadedFlightPlan.allLegs
-        .map((l, index) =>
-          l.isDiscontinuity === false && index >= activeLegIndex && l.cruiseStep ? l.cruiseStep : null,
-        )
-        .filter((it) => it !== null);
-      const cruiseStepLegIndices = this.loadedFlightPlan.allLegs
-        .map((l, index) => (l.isDiscontinuity === false && index >= activeLegIndex && l.cruiseStep ? index : null))
-        .filter((it) => it !== null);
+    if (
+      this.loadedFlightPlan &&
+      this.selectedPageIndex.get() === SelectedPage.STEP_ALTS &&
+      !this.selectedLegIsAlternate.get()
+    ) {
+      const stepDisabledReason = this.checkStepAltsAccessPrerequisites();
+      this.stepPageDisabledReason.set(stepDisabledReason ?? null);
+      if (stepDisabledReason === undefined) {
+        const activeLegIndex = this.loadedFlightPlan.activeLegIndex;
+        const cruiseSteps = this.loadedFlightPlan.allLegs
+          .map((l, index) =>
+            l.isDiscontinuity === false && index >= activeLegIndex && l.cruiseStep ? l.cruiseStep : null,
+          )
+          .filter((it) => it !== null);
+        const cruiseStepLegIndices = this.loadedFlightPlan.allLegs
+          .map((l, index) => (l.isDiscontinuity === false && index >= activeLegIndex && l.cruiseStep ? index : null))
+          .filter((it) => it !== null);
 
-      for (let i = 0; i < 5; i++) {
-        this.stepAltsLineVisibility[i].set('hidden');
-        this.stepLinesPredictionsVisibility[i].set('hidden');
-      }
+        for (let i = 0; i < 5; i++) {
+          this.stepAltsLineVisibility[i].set('hidden');
+          this.stepLinesPredictionsVisibility[i].set('hidden');
+        }
 
-      if (this.stepAltsNumberOfCruiseSteps.get() !== cruiseSteps.length) {
-        this.forceRebuildList = true;
-      }
+        if (this.stepAltsNumberOfCruiseSteps.get() !== cruiseSteps.length) {
+          this.forceRebuildList = true;
+        }
 
-      this.stepAltsNumberOfCruiseSteps.set(cruiseSteps.length);
-      this.stepAltsScrollUpDisabled.set(this.stepAltsStartAtStepIndex.get() === 0);
-      this.stepAltsScrollDownDisabled.set(cruiseSteps.length - this.stepAltsStartAtStepIndex.get() <= 4);
+        this.stepAltsNumberOfCruiseSteps.set(cruiseSteps.length);
+        this.stepAltsScrollUpDisabled.set(this.stepAltsStartAtStepIndex.get() === 0);
+        this.stepAltsScrollDownDisabled.set(cruiseSteps.length - this.stepAltsStartAtStepIndex.get() <= 4);
 
-      for (let i = this.stepAltsStartAtStepIndex.get(); i < this.stepAltsStartAtStepIndex.get() + 5; i++) {
-        const line = i - this.stepAltsStartAtStepIndex.get();
+        for (let i = this.stepAltsStartAtStepIndex.get(); i < this.stepAltsStartAtStepIndex.get() + 5; i++) {
+          const line = i - this.stepAltsStartAtStepIndex.get();
 
-        if (!(i in cruiseSteps)) {
-          if (this.forceRebuildList) {
-            this.stepAltsWptIndices[line].set(null);
-            this.stepAltsWptIndices[line].notify();
-            this.stepAltsAltitude[line].set(null);
-            this.forceRebuildList = false;
+          if (!(i in cruiseSteps)) {
+            if (this.forceRebuildList) {
+              this.stepAltsWptIndices[line].set(null);
+              this.stepAltsWptIndices[line].notify();
+              this.stepAltsAltitude[line].set(null);
+              this.forceRebuildList = false;
+            }
+            this.stepLinesPredictionsVisibility[line].set('hidden');
+            this.stepAltsLineVisibility[line].set('visible');
+            this.stepAltsDistances[line].set(null);
+            this.stepAltsTimes[line].set('--:--');
+            this.stepAltsIgnored[line].set(false);
+            this.stepAltsAboveMaxFl[line].set(false);
+
+            break;
           }
-          this.stepLinesPredictionsVisibility[line].set('hidden');
+
+          const pred =
+            this.loadedFlightPlanIndex.get() < FlightPlanIndex.Uplink
+              ? this.props.fmcService?.master?.guidanceController?.vnavDriver?.mcduProfile?.waypointPredictions?.get(
+                  cruiseStepLegIndices[i],
+                )
+              : undefined;
+          const wptEta = getEtaFromUtcOrPresent(
+            pred?.secondsFromPresent,
+            this.activeFlightPhase.get() == FmgcFlightPhase.Preflight,
+          );
+          const wptIndex = this.availableWaypointsToLegIndex.indexOf(cruiseStepLegIndices[i]);
+          const step = cruiseSteps[i];
+
           this.stepAltsLineVisibility[line].set('visible');
-          this.stepAltsDistances[line].set(null);
-          this.stepAltsTimes[line].set('--:--');
-          this.stepAltsIgnored[line].set(false);
-          this.stepAltsAboveMaxFl[line].set(false);
+          this.stepAltsWptIndices[line].set(wptIndex !== -1 ? wptIndex : null);
+          this.stepAltsAltitude[line].set(step.toAltitude);
+          this.stepLinesPredictionsVisibility[line].set(
+            wptIndex !== -1 && step.toAltitude !== null ? 'visible' : 'hidden',
+          );
 
-          break;
+          if (pred) {
+            this.stepAltsDistances[line].set(pred.distanceFromAircraft - step.distanceBeforeTermination);
+            this.stepAltsTimes[line].set(wptEta);
+          } else {
+            this.stepAltsDistances[line].set(null);
+            this.stepAltsTimes[line].set('--:--');
+          }
+          this.stepAltsIgnored[line].set(step.isIgnored);
+
+          const estGrossWeight = this.getEstimatedGrossWeightAtIndex(cruiseStepLegIndices[i]);
+          this.stepAltsAboveMaxFl[line].set(
+            estGrossWeight !== null
+              ? step.toAltitude >
+                  (this.props.fmcService.master.getRecMaxAltitude(this.loadedFlightPlanIndex.get(), estGrossWeight) ??
+                    Infinity)
+              : false,
+          );
+
+          if (this.stepAltsIgnored[line].get()) {
+            this.stepAltsMessage[line].set('IGNORED');
+          } else if (this.stepAltsAboveMaxFl[line].get()) {
+            this.stepAltsMessage[line].set('ABOVE MAX FL');
+          } else {
+            this.stepAltsMessage[line].set('');
+          }
         }
+      }
+    }
+  }
 
-        const pred =
-          this.loadedFlightPlanIndex.get() < FlightPlanIndex.Uplink
-            ? this.props.fmcService?.master?.guidanceController?.vnavDriver?.mcduProfile?.waypointPredictions?.get(
-                cruiseStepLegIndices[i],
-              )
-            : undefined;
-        const wptEta = getEtaFromUtcOrPresent(
-          pred?.secondsFromPresent,
-          this.activeFlightPhase.get() == FmgcFlightPhase.Preflight,
-        );
-        const wptIndex = this.availableWaypointsToLegIndex.indexOf(cruiseStepLegIndices[i]);
-        const step = cruiseSteps[i];
-
-        this.stepAltsLineVisibility[line].set('visible');
-        this.stepAltsWptIndices[line].set(wptIndex !== -1 ? wptIndex : null);
-        this.stepAltsAltitude[line].set(step.toAltitude);
-        this.stepLinesPredictionsVisibility[line].set(
-          wptIndex !== -1 && step.toAltitude !== null ? 'visible' : 'hidden',
-        );
-
-        if (pred) {
-          this.stepAltsDistances[line].set(pred.distanceFromAircraft - step.distanceBeforeTermination);
-          this.stepAltsTimes[line].set(wptEta);
-        } else {
-          this.stepAltsDistances[line].set(null);
-          this.stepAltsTimes[line].set('--:--');
-        }
-        this.stepAltsIgnored[line].set(step.isIgnored);
-
-        const estGrossWeight = this.getEstimatedGrossWeightAtIndex(cruiseStepLegIndices[i]);
-        this.stepAltsAboveMaxFl[line].set(
-          estGrossWeight !== null
-            ? step.toAltitude >
-                (this.props.fmcService.master.getRecMaxAltitude(this.loadedFlightPlanIndex.get(), estGrossWeight) ??
-                  Infinity)
-            : false,
-        );
-
-        if (this.stepAltsIgnored[line].get()) {
-          this.stepAltsMessage[line].set('IGNORED');
-        } else if (this.stepAltsAboveMaxFl[line].get()) {
-          this.stepAltsMessage[line].set('ABOVE MAX FL');
-        } else {
-          this.stepAltsMessage[line].set('');
-        }
+  /**
+   * Checks if step altitude modification is allowed based on flight phase, engine out status and whether a cruise flight level is set.
+   * @returns the reason why the step altitude modification is not allowed, or undefined if it is allowed.
+   */
+  private checkStepAltsAccessPrerequisites(): StepDisabledReason | undefined {
+    if (this.props.fmcService.master.engineOutActive()) {
+      return StepDisabledReason.EngineOut;
+    } else {
+      const fp = this.props.flightPlanInterface.get(this.loadedFlightPlanIndex.get());
+      const phase = fp.isActiveOrCopiedFromActive() ? this.activeFlightPhase.get() : null;
+      if (phase === FmgcFlightPhase.Descent) {
+        return StepDisabledReason.Descent;
+      } else if (phase === FmgcFlightPhase.Approach) {
+        return StepDisabledReason.Approach;
+      } else if (phase === FmgcFlightPhase.GoAround) {
+        return StepDisabledReason.GoAround;
+      } else if (phase === FmgcFlightPhase.Done) {
+        return StepDisabledReason.Done;
+      } else if (fp.performanceData.cruiseFlightLevel.get() === null) {
+        return StepDisabledReason.NoCruiseLevel;
       }
     }
   }
@@ -548,29 +637,29 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
   }
 
   private checkLegModificationAllowed(): boolean {
-    return this.props.fmcService.master !== null && this.selectedLegIndex !== null && this.loadedFlightPlan !== null;
+    return (
+      this.props.fmcService.master !== null && this.selectedLegIndex.get() !== null && this.loadedFlightPlan !== null
+    );
   }
 
   private async onWptDropdownModified(idx: number | null): Promise<void> {
     if (idx !== null) {
       const legIndex = this.availableWaypointsToLegIndex[idx];
-      this.selectedLegIndex = legIndex;
-      this.props.fmcService.master.revisedLegIndex.set(legIndex);
+      this.selectedLegIndex.set(legIndex);
       this.updateConstraints();
     } else {
-      this.selectedLegIndex = null;
-      this.props.fmcService.master.resetRevisedWaypoint();
+      this.selectedLegIndex.set(null);
     }
   }
 
   private async tryUpdateSpeedConstraint(newSpeed?: number) {
     if (this.checkLegModificationAllowed() && this.spdConstraintTypeRadioSelected.get() !== null) {
       this.props.fmcService.master!.flightPlanInterface.setPilotEnteredSpeedConstraintAt(
-        this.selectedLegIndex!,
+        this.selectedLegIndex.get()!,
         this.spdConstraintTypeRadioSelected.get() === 1,
         newSpeed,
         this.loadedFlightPlanIndex.get(),
-        false,
+        this.selectedLegIsAlternate.get() ?? false,
       );
     }
   }
@@ -605,11 +694,11 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
       }
 
       this.props.fmcService.master!.flightPlanInterface.setPilotEnteredAltitudeConstraintAt(
-        this.selectedLegIndex!,
+        this.selectedLegIndex.get()!,
         this.altConstraintTypeRadioSelected.get() === 1,
         { altitude1: alt, altitudeDescriptor: option },
         this.loadedFlightPlanIndex.get(),
-        false,
+        this.selectedLegIsAlternate.get() ?? false,
       );
     }
   }
@@ -617,11 +706,11 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
   private deleteAltitudeConstraint() {
     if (this.checkLegModificationAllowed()) {
       this.props.fmcService.master!.flightPlanInterface.setPilotEnteredAltitudeConstraintAt(
-        this.selectedLegIndex!,
+        this.selectedLegIndex.get()!,
         this.altConstraintTypeRadioSelected.get() === 1,
         undefined,
         this.loadedFlightPlanIndex.get(),
-        false,
+        this.selectedLegIsAlternate.get() ?? false,
       );
     }
   }
@@ -638,13 +727,13 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
         this.props.flightPlanInterface.setPilotEntryClimbSpeedLimitSpeed(
           value,
           this.loadedFlightPlanIndex.get(),
-          false,
+          this.selectedLegIsAlternate.get() ?? false,
         );
       } else {
         this.props.flightPlanInterface.setPilotEntryDescentSpeedLimitSpeed(
           value,
           this.loadedFlightPlanIndex.get(),
-          false,
+          this.selectedLegIsAlternate.get() ?? false,
         );
       }
     }
@@ -658,13 +747,13 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
         this.props.flightPlanInterface.setPilotEntryClimbSpeedLimitAltitude(
           value,
           this.loadedFlightPlanIndex.get(),
-          false,
+          this.selectedLegIsAlternate.get() ?? false,
         );
       } else {
         this.props.flightPlanInterface.setPilotEntryDescentSpeedLimitAltitude(
           value,
           this.loadedFlightPlanIndex.get(),
-          false,
+          this.selectedLegIsAlternate.get() ?? false,
         );
       }
     }
@@ -673,9 +762,15 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
   private async deleteSpeedLimit() {
     if (this.checkPerformanceDataEditCondition()) {
       if (this.speedLimitType.get() === SpeedLimitType.CLB) {
-        this.props.flightPlanInterface.deleteClimbSpeedLimit(this.loadedFlightPlanIndex.get(), false);
+        this.props.flightPlanInterface.deleteClimbSpeedLimit(
+          this.loadedFlightPlanIndex.get(),
+          this.selectedLegIsAlternate.get() ?? false,
+        );
       } else {
-        this.props.flightPlanInterface.deleteDescentSpeedLimit(this.loadedFlightPlanIndex.get(), false);
+        this.props.flightPlanInterface.deleteDescentSpeedLimit(
+          this.loadedFlightPlanIndex.get(),
+          this.selectedLegIsAlternate.get() ?? false,
+        );
       }
     }
   }
@@ -767,8 +862,9 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
       if (isValid) {
         this.props.flightPlanInterface.addOrUpdateCruiseStep(legIndex, altitude, this.loadedFlightPlanIndex.get());
       } else {
-        if (this.selectedLegIndex !== null) {
-          const leg = this.loadedFlightPlan?.maybeElementAt(this.selectedLegIndex);
+        const selectedLegIndex = this.selectedLegIndex.get();
+        if (selectedLegIndex !== null) {
+          const leg = this.loadedFlightPlan?.maybeElementAt(selectedLegIndex);
           this.stepNotAllowedAt.set(leg?.isDiscontinuity === false ? `STEP NOT ALLOWED AT ${leg.ident}` : '');
         }
       }
@@ -838,18 +934,24 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
     // If extra parameter for activeUri is given, navigate to flight phase sub-page
     switch (this.props.mfd.uiService.activeUri.get().extra) {
       case 'rta':
+        this.initialLoadRevisedWaypointIndex = false;
         this.selectedPageIndex.set(SelectedPage.RTA);
         break;
       case 'spd':
+        this.selectedLegIsAlternate.set(this.props.fmcService.master.revisedLegIsAltn.get());
+        this.initialLoadRevisedWaypointIndex = true;
         this.selectedPageIndex.set(SelectedPage.SPD);
         break;
       case 'cms':
         this.selectedPageIndex.set(SelectedPage.CMS);
         break;
       case 'alt':
+        this.selectedLegIsAlternate.set(this.props.fmcService.master.revisedLegIsAltn.get());
+        this.initialLoadRevisedWaypointIndex = true;
         this.selectedPageIndex.set(SelectedPage.ALT);
         break;
       case 'step-alts':
+        this.initialLoadRevisedWaypointIndex = false;
         this.selectedPageIndex.set(SelectedPage.STEP_ALTS);
         break;
 
@@ -899,12 +1001,13 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
     );
 
     this.subs.push(
-      this.spdConstraintTypeRadioSelected.sub(() => this.tryUpdateSpeedConstraint()),
-      this.altConstraintTypeRadioSelected.sub(() => this.tryUpdateAltitudeConstraint()),
-    );
-
-    this.subs.push(
       this.selectedPageIndex.sub((val) => {
+        if (val === SelectedPage.STEP_ALTS || val === SelectedPage.RTA || val === SelectedPage.CMS) {
+          this.selectedLegIsAlternate.set(false);
+        } else if (val === SelectedPage.SPD || val === SelectedPage.ALT) {
+          this.selectedLegIsAlternate.set(this.props.fmcService.master.revisedLegIsAltn.get());
+        }
+
         if (val === SelectedPage.STEP_ALTS) {
           this.stepsAltsClockSub?.resume();
         } else {
@@ -912,6 +1015,9 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
         }
         this.onNewData();
       }, true),
+      this.stepPageDisabledMessageDisplay,
+      this.stepAltsPageDisplay,
+      this.fillPageDisplay,
     );
   }
 
@@ -999,6 +1105,9 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
                         selectedIndex={this.spdConstraintTypeRadioSelected}
                         values={['CLB CSTR', 'DES CSTR']}
                         color={Subject.create(RadioButtonColor.Amber)}
+                        onModified={() => {
+                          this.tryUpdateSpeedConstraint();
+                        }}
                       />
                     </div>
                     <Button
@@ -1168,6 +1277,9 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
                           selectedIndex={this.altConstraintTypeRadioSelected}
                           values={['CLB CSTR', 'DES CSTR']}
                           color={Subject.create(RadioButtonColor.Amber)}
+                          onModified={() => {
+                            this.tryUpdateAltitudeConstraint();
+                          }}
                         />
                       </div>
                     </div>
@@ -1176,7 +1288,18 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
               </TopTabNavigatorPage>
               <TopTabNavigatorPage>
                 {/* STEP ALTs */}
-                <div style="flex: 1; display: flex; flex-direction: column; justify-content: center; align-items: center;">
+                <div class="mfd-page-container" style={{ display: this.stepPageDisabledMessageDisplay }}>
+                  <div class="mfd-amber-error-message">{this.stepPageDisabledReason}</div>
+                </div>
+                <div
+                  style={{
+                    flex: '1',
+                    display: this.stepAltsPageDisplay,
+                    flexDirection: 'column',
+                    justifyContent: 'center',
+                    alignItems: 'center',
+                  }}
+                >
                   <div class="mfd-fms-fpln-labeled-box-container" style="width: 100%;">
                     <div class="mfd-fms-fpln-labeled-box-label" style="margin-left: 15px;">
                       <span class="mfd-label mfd-spacing-right">STEP ALTs FROM CRZ</span>
@@ -1333,7 +1456,7 @@ export class MfdFmsFplnVertRev extends FmsPage<MfdFmsFplnVertRevProps> {
                 </div>
               </TopTabNavigatorPage>
             </TopTabNavigator>
-            <div style="flex-grow: 1;" />
+            <div style={{ 'flex-grow': '1', display: this.fillPageDisplay }} />
             <div style="display: flex; flex-direction: row; justify-content: space-between;">
               <div ref={this.returnButtonDiv} style="display: flex; justify-content: flex-end; padding: 2px;">
                 <Button