+++ fbw-a380x/src/systems/instruments/src/MFD/pages/FMS/MfdFmsPerf.tsx
@@ -94,6 +94,8 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
 
   private readonly flightPlanChangeNotifier = new FlightPlanChangeNotifier(this.props.bus);
 
+  private readonly isActiveOrCopyOfActive = Subject.create(false);
+
   private readonly destEfobAmber = MappedSubject.create(
     ([destEfobBelowM, loadedFpIndex]) => destEfobBelowM && loadedFpIndex === FlightPlanIndex.Active,
     this.props.fmcService.master.fmgc.data.destEfobBelowMinInActive,
@@ -213,45 +215,52 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
   }
 
   private readonly toPageInactive = MappedSubject.create(
-    ([flightPhase, flightPlan]) => flightPlan === FlightPlanIndex.Active && flightPhase >= FmgcFlightPhase.Takeoff,
+    ([flightPhase, isActiveOrCopyOfActive]) => isActiveOrCopyOfActive && flightPhase >= FmgcFlightPhase.Takeoff,
     this.activeFlightPhase,
-    this.loadedFlightPlanIndex,
+    this.isActiveOrCopyOfActive,
   );
   private readonly clbPageInactive = MappedSubject.create(
-    ([flightPhase, flightPlan]) => flightPlan === FlightPlanIndex.Active && flightPhase >= FmgcFlightPhase.Climb,
+    ([flightPhase, isActiveOrCopyOfActive]) => isActiveOrCopyOfActive && flightPhase > FmgcFlightPhase.Climb,
     this.activeFlightPhase,
-    this.loadedFlightPlanIndex,
+    this.isActiveOrCopyOfActive,
   );
-  private readonly crzPageInactive = MappedSubject.create(
-    ([flightPhase, flightPlan]) => flightPlan === FlightPlanIndex.Active && flightPhase >= FmgcFlightPhase.Cruise,
+
+  private readonly atOrAfterClimbPhase = MappedSubject.create(
+    ([flightPhase, isActiveOrCopyOfActive]) => isActiveOrCopyOfActive && flightPhase >= FmgcFlightPhase.Climb,
     this.activeFlightPhase,
-    this.loadedFlightPlanIndex,
+    this.isActiveOrCopyOfActive,
+  );
+
+  private readonly crzPageInactive = MappedSubject.create(
+    ([flightPhase, isActiveOrCopyOfActive]) => isActiveOrCopyOfActive && flightPhase > FmgcFlightPhase.Cruise,
+    this.activeFlightPhase,
+    this.isActiveOrCopyOfActive,
   );
   private readonly desPageInactive = MappedSubject.create(
-    ([flightPhase, flightPlan]) => flightPlan === FlightPlanIndex.Active && flightPhase >= FmgcFlightPhase.Descent,
+    ([flightPhase, isActiveOrCopyOfActive]) => isActiveOrCopyOfActive && flightPhase > FmgcFlightPhase.Descent,
     this.activeFlightPhase,
-    this.loadedFlightPlanIndex,
+    this.isActiveOrCopyOfActive,
   );
 
   private readonly notYetInClimb = MappedSubject.create(
-    ([flightPhase, flightPlan]) => flightPlan !== FlightPlanIndex.Active || flightPhase < FmgcFlightPhase.Climb,
+    ([flightPhase, isActiveOrCopyOfActive]) => !isActiveOrCopyOfActive || flightPhase < FmgcFlightPhase.Climb,
     this.activeFlightPhase,
-    this.loadedFlightPlanIndex,
+    this.isActiveOrCopyOfActive,
   );
   private readonly notYetInCruise = MappedSubject.create(
-    ([flightPhase, flightPlan]) => flightPlan !== FlightPlanIndex.Active || flightPhase < FmgcFlightPhase.Cruise,
+    ([flightPhase, isActiveOrCopyOfActive]) => !isActiveOrCopyOfActive || flightPhase < FmgcFlightPhase.Cruise,
     this.activeFlightPhase,
-    this.loadedFlightPlanIndex,
+    this.isActiveOrCopyOfActive,
   );
   private readonly notYetInDescent = MappedSubject.create(
-    ([flightPhase, flightPlan]) => flightPlan !== FlightPlanIndex.Active || flightPhase < FmgcFlightPhase.Descent,
+    ([flightPhase, isActiveOrCopyOfActive]) => !isActiveOrCopyOfActive || flightPhase < FmgcFlightPhase.Descent,
     this.activeFlightPhase,
-    this.loadedFlightPlanIndex,
+    this.isActiveOrCopyOfActive,
   );
   private readonly notInDescent = MappedSubject.create(
-    ([flightPhase, flightPlan]) => flightPlan !== FlightPlanIndex.Active || flightPhase !== FmgcFlightPhase.Descent,
+    ([flightPhase, isActiveOrCopyOfActive]) => !isActiveOrCopyOfActive || flightPhase !== FmgcFlightPhase.Descent,
     this.activeFlightPhase,
-    this.loadedFlightPlanIndex,
+    this.isActiveOrCopyOfActive,
   );
 
   // TO page subjects, refs and methods
@@ -427,21 +436,17 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
 
   private readonly costIndexModeLabels = ArraySubject.create(['LRC', 'ECON']);
   private readonly costIndexDisabled = MappedSubject.create(
-    ([flightPhase, ciMode, fpIndex]) =>
-      ciMode == CostIndexMode.LRC ||
-      (flightPhase >= FmgcFlightPhase.Descent &&
-        this.props.flightPlanInterface.get(fpIndex).isActiveOrCopiedFromActive()),
+    ([flightPhase, ciMode, isActiveOrCopyOfActive]) =>
+      ciMode == CostIndexMode.LRC || (flightPhase >= FmgcFlightPhase.Descent && isActiveOrCopyOfActive),
     this.activeFlightPhase,
     this.costIndexMode,
-    this.loadedFlightPlanIndex,
+    this.isActiveOrCopyOfActive,
   );
 
   private readonly costIndexModeDisabled = MappedSubject.create(
-    ([flightPhase, fpIndex]) =>
-      this.props.flightPlanInterface.get(fpIndex).isActiveOrCopiedFromActive() &&
-      flightPhase >= FmgcFlightPhase.Descent,
+    ([flightPhase, isActiveOrCopyOfActive]) => isActiveOrCopyOfActive && flightPhase >= FmgcFlightPhase.Descent,
     this.activeFlightPhase,
-    this.loadedFlightPlanIndex,
+    this.isActiveOrCopyOfActive,
   );
 
   private readonly speedConstraintSpeed = Subject.create<number | null>(null);
@@ -748,6 +753,8 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
 
     const fpIndex = this.loadedFlightPlanIndex.get();
 
+    this.isActiveOrCopyOfActive.set(this.props.flightPlanInterface.get(fpIndex).isActiveOrCopiedFromActive());
+
     this.crzFlIsMandatory.set(
       fpIndex === FlightPlanIndex.Active &&
         (this.props.fmcService.master.fmgc.getFlightPhase() ?? FmgcFlightPhase.Preflight) < FmgcFlightPhase.Descent,
@@ -887,10 +894,7 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
       this.activeFlightPhase.sub((val) => {
         if (this.previousFmsFlightPhase) {
           const isSamePhase = this.flightPhasesSelectedPageIndex.get() + 1 === this.previousFmsFlightPhase;
-          if (
-            isSamePhase &&
-            this.props.flightPlanInterface.get(this.loadedFlightPlanIndex.get()).isActiveOrCopiedFromActive()
-          ) {
+          if (isSamePhase && this.isActiveOrCopyOfActive.get()) {
             switch (val) {
               case FmgcFlightPhase.Takeoff:
               case FmgcFlightPhase.Climb:
@@ -1004,7 +1008,8 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
       this.destEfobAmber,
       this.flightPhasesSelectedPageIndex.sub((val) => this.drawPage(val)),
       this.costIndexDisabled,
-      this.costIndexDisabled,
+      this.costIndexModeDisabled,
+      this.atOrAfterClimbPhase,
     );
   }
 
@@ -2169,7 +2174,7 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                       componentIfTrue={
                         <InputField<number>
                           dataEntryFormat={new SpeedKnotsFormat(Subject.create(90), Subject.create(Vmo))}
-                          inactive={this.clbPageInactive}
+                          inactive={this.atOrAfterClimbPhase}
                           value={this.climbPreselectedSpeed}
                           dataHandlerDuringValidation={(v) =>
                             this.props.flightPlanInterface.setPerformanceData(
@@ -2251,7 +2256,7 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                   <div style="margin-bottom: 15px;">
                     <InputField<number>
                       dataEntryFormat={new AltitudeOrFlightLevelFormat(this.transAlt)}
-                      inactive={this.clbPageInactive}
+                      inactive={this.atOrAfterClimbPhase}
                       enteredByPilot={this.thrRedAltIsPilotEntered}
                       value={this.thrRedAlt}
                       dataHandlerDuringValidation={async (v) =>
@@ -2293,7 +2298,7 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                             this.loadedFlightPlanIndex.get(),
                           );
                         }}
-                        inactive={this.clbPageInactive}
+                        inactive={this.atOrAfterClimbPhase}
                         readonlyValue={this.noiseN1}
                         containerStyle="width: 110px;"
                         alignText="flex-end"
@@ -2339,7 +2344,7 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                   <div style="margin-bottom: 15px;">
                     <InputField<number>
                       dataEntryFormat={new AltitudeOrFlightLevelFormat(this.transAlt)}
-                      inactive={this.clbPageInactive}
+                      inactive={this.atOrAfterClimbPhase}
                       enteredByPilot={this.accelAltIsPilotEntered}
                       value={this.accelAlt}
                       dataHandlerDuringValidation={async (v) =>
@@ -2381,7 +2386,7 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                             this.loadedFlightPlanIndex.get(),
                           );
                         }}
-                        inactive={this.clbPageInactive}
+                        inactive={this.atOrAfterClimbPhase}
                         readonlyValue={this.noiseSpeed}
                         containerStyle="width: 110px;"
                         alignText="flex-end"
@@ -2427,7 +2432,7 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                             this.loadedFlightPlanIndex.get(),
                           );
                         }}
-                        inactive={this.clbPageInactive}
+                        inactive={this.atOrAfterClimbPhase}
                         readonlyValue={this.noiseEndAltitude}
                         containerStyle="width: 150px;"
                         alignText="flex-end"
@@ -2486,26 +2491,15 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                       interactionMode={this.props.mfd.interactionMode}
                     />
                   </div>
-                  <div>
-                    <ConditionalComponent
-                      width={140}
-                      height={40}
-                      condition={this.toPageInactive}
-                      componentIfFalse={
-                        <Button
-                          label="SPD CSTR"
-                          onClick={() =>
-                            this.props.mfd.uiService.navigateTo(
-                              `fms/${this.props.mfd.uiService.activeUri.get().category}/f-pln-vert-rev/spd`,
-                            )
-                          }
-                        >
-                          SPD CSTR
-                        </Button>
-                      }
-                      componentIfTrue={<></>}
-                    />
-                  </div>
+                  <Button
+                    label="SPD CSTR"
+                    onClick={() =>
+                      this.props.mfd.uiService.navigateTo(
+                        `fms/${this.props.mfd.uiService.activeUri.get().category}/f-pln-vert-rev/spd`,
+                      )
+                    }
+                    disabled={this.clbPageInactive}
+                  ></Button>
                 </div>
               </TopTabNavigatorPage>
               <TopTabNavigatorPage containerStyle="padding-top: 0px; padding-left: 0px;">
@@ -2623,7 +2617,6 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                       componentIfTrue={
                         <InputField<number, number, false>
                           dataEntryFormat={new SpeedMachFormat(Subject.create(0.1), Subject.create(Mmo))}
-                          inactive={this.crzPageInactive}
                           onModified={async (v) => {
                             this.props.flightPlanInterface.setPerformanceData(
                               'preselectedCruiseSpeed',
@@ -2651,7 +2644,6 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                       componentIfTrue={
                         <InputField<number, number, false>
                           dataEntryFormat={new SpeedKnotsFormat(Subject.create(90), Subject.create(Vmo))}
-                          inactive={this.crzPageInactive}
                           onModified={async (v) => {
                             this.props.flightPlanInterface.setPerformanceData(
                               'preselectedCruiseSpeed',
@@ -2802,6 +2794,7 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                           `fms/${this.props.mfd.uiService.activeUri.get().category}/f-pln-vert-rev/step-alts`,
                         )
                       }
+                      disabled={this.crzPageInactive}
                     />
                   </div>
                 </div>
@@ -2911,7 +2904,7 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                   </div>
                   <div class="mfd-fms-perf-speed-table-cell">
                     <ConditionalComponent
-                      condition={this.desPageInactive}
+                      condition={this.notYetInDescent}
                       componentIfFalse={
                         <InputField<number>
                           dataEntryFormat={new SpeedMachFormat(Subject.create(0.1), Subject.create(Mmo))}
@@ -2922,7 +2915,6 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                               this.loadedFlightPlanIndex.get(),
                             );
                           }}
-                          inactive={this.desPageInactive}
                           value={this.desManagedMachTarget}
                           alignText="flex-end"
                           errorHandler={(e) => this.props.fmcService.master.showFmsErrorMessage(e.type, e.details)}
@@ -2950,7 +2942,6 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                               this.loadedFlightPlanIndex.get(),
                             );
                           }}
-                          inactive={this.desPageInactive}
                           value={this.desManagedSpdTarget}
                           alignText="flex-end"
                           errorHandler={(e) => this.props.fmcService.master.showFmsErrorMessage(e.type, e.details)}
@@ -3030,6 +3021,7 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                           `fms/${this.props.mfd.uiService.activeUri.get().category}/f-pln-vert-rev/spd`,
                         )
                       }
+                      disabled={this.desPageInactive}
                     />
                   </div>
                 </div>